+++ src/Test/Integration/Model/Service/ImportOrderTest.php
@@ -219,6 +219,67 @@ class ImportOrderTest extends TestCase
     }
 
     /**
+     * @param string $expectedState
+     * @param string $country
+     * @param string $state
+     *
+     * @dataProvider regionProvider
+     * @throws ShopgateLibraryException
+     *
+     * @magentoConfigFixture default/general/region/state_required BZ
+     */
+    public function testRequiredStates($expectedState, $country, $state): void
+    {
+        $invoiceAddress            = $this->dataManager->getGermanAddress();
+        $invoiceAddress['country'] = $country;
+        $invoiceAddress['state']   = $state;
+        $billingAddress            = $this->dataManager->getGermanAddress(false);
+        $billingAddress['country'] = $country;
+        $billingAddress['state']   = $state;
+
+        $shopgateOrder = new \ShopgateOrder(
+            [
+                'order_number'               => random_int(1000000000, 9999999999),
+                'is_paid'                    => 1,
+                'payment_time'               => null,
+                'payment_transaction_number' => '8654415',
+                'mail'                       => 'shopgate@shopgate.com',
+                'amount_shop_payment'        => '5.00',
+                'amount_complete'            => '149.85',
+                'shipping_infos'             => ['amount' => '4.90'],
+                'invoice_address'            => $invoiceAddress,
+                'delivery_address'           => $billingAddress,
+                'external_coupons'           => [],
+                'shopgate_coupons'           => [],
+                'items'                      => [$this->dataManager->getSimpleProduct()],
+                'payment_method'             => 'COD',
+                'payment_group'              => 'COD'
+            ]
+        );
+
+        $this->orderHolder[] = $this->importClass->addOrder($shopgateOrder);
+        /** @var ShopgateOrder $sgOrder */
+        $sgOrder = Bootstrap::getObjectManager()->get(ShopgateOrder::class);
+        /** @var MageOrder $order */
+        $order = $sgOrder->loadMethods([]);
+
+        $this->assertEquals($expectedState, $order->getShippingAddress()->getRegion());
+    }
+
+    /**
+     * @return array
+     */
+    public function regionProvider(): array
+    {
+        return [
+            'Mapped region'                      => ['Sachsen-Anhalt', 'DE', 'DE-ST'],
+            'Not a mapped region'                => ['TOL', 'BZ', 'BZ-TOL'],
+            'No region required'                 => ['', 'DE', ''],
+            'Should not fail, although required' => ['', 'BZ', ''],
+        ];
+    }
+
+    /**
      * @return array
      * @throws Exception
      */